refactor: merge paypal standard gateways
