+++ src/PaymentGateways/Gateways/PayPalStandard/resources/js/PayPalLogo.tsx
