+++ src/PaymentGateways/Gateways/PayPalStandard/resources/js/payPalStandardGateway.tsx
